03_django_redis_celery/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.7
django-celery-beat==2.6.0
django-timezone-field==7.0
gunicorn==22.0.0
kombu==5.3.7
packaging==24.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
redis==5.0.7
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
